backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
yfinance>=0.2.40
cachetools>=5.5.0
python-dateutil>=2.9.0
pydantic>=2.0
requests>=2.32.0
